refactor: 💡 Move generated `TxTags` to `generated/types.ts`
