+++ package.json
@@ -8,15 +8,14 @@
   "license": "ISC",
   "scripts": {
     "start": "cross-env node scripts/generateTsconfigDev.js && cross-env webpack-dev-server --config=webpack.config.dev.js",
-    "generate:polkadot-types": "yarn fetch-metadata && yarn fetch-definitions && yarn generate:defs && yarn generate:meta && yarn generate:tags && yarn generate:country-codes && yarn generate:post-process",
+    "generate:polkadot-types": "yarn fetch-metadata && yarn fetch-definitions && yarn generate:defs && yarn generate:meta && yarn generate:consts && yarn generate:post-process",
     "generate:middleware-types": "graphql-codegen -r dotenv/config",
     "fetch-definitions": "cross-env node scripts/fetchDefinitions.js",
     "fetch-metadata": "sh scripts/fetchMetadata.sh",
     "update-readme": "cross-env node scripts/updateReadme.js",
     "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --endpoint metadata.json --package polymesh-types --input ./src/polkadot/",
     "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --endpoint metadata.json --package polymesh-types --output ./src/polkadot --strict",
-    "generate:tags": "cross-env node scripts/generateTxTags.js",
-    "generate:country-codes": "cross-env node scripts/generateCountryCodes.js",
+    "generate:consts": "cross-env node scripts/generateConsts.js",
     "generate:post-process": "cross-env node scripts/postProcessTypes.js",
     "test": "jest --detectOpenHandles --coverage",
     "test:no-cov": "jest --detectOpenHandles",
