+++ src/types/internal.ts
@@ -13,7 +13,7 @@ import { DocumentNode } from 'graphql';
 
 import { Identity, PostTransactionValue } from '~/internal';
 import { CallIdEnum, ModuleIdEnum } from '~/middleware/types';
-import { CustomAssetTypeId, TxTag } from '~/polkadot';
+import { CustomAssetTypeId } from '~/polkadot';
 import {
   CalendarPeriod,
   KnownAssetType,
@@ -21,6 +21,7 @@ import {
   Role,
   SignerValue,
   SimplePermissions,
+  TxTag,
 } from '~/types';
 
 /**
